+++ livestreams.go
@@ -214,7 +214,7 @@ func (s *LiveStreams) doRequest(
 // Suppose a content creator is hosting a live concert and wants to display the number of live viewers on
 // their dashboard. This endpoint can be queried to show up-to-date viewer statistics.
 //
-// Related guide: <a href="https://fastpix.com/docs/manage-live-streams/create-and-manage-live-streams">Manage streams</a>
+// Related guide: <a href="https://fastpix.com/docs/live-streaming/create-and-manage-live-streams">Manage streams</a>
 func (s *LiveStreams) GetViewerCount(ctx context.Context, streamID string, opts ...operations.Option) (*operations.GetLiveStreamViewerCountByIDResponse, error) {
 	request := operations.GetLiveStreamViewerCountByIDRequest{StreamID: streamID}
 
@@ -301,7 +301,7 @@ func (s *LiveStreams) parseGetViewerCountResponse(req *http.Request, httpRes *ht
 // Suppose a news agency is broadcasting a live event and wants to track the configurations set for the
 // live stream while also checking the stream's status.
 //
-// Related guide: <a href="https://fastpix.com/docs/manage-live-streams/create-and-manage-live-streams">Manage streams</a>
+// Related guide: <a href="https://fastpix.com/docs/live-streaming/create-and-manage-live-streams">Manage streams</a>
 func (s *LiveStreams) GetByID(ctx context.Context, streamID string, opts ...operations.Option) (*operations.GetLiveStreamByIDResponse, error) {
 	request := operations.GetLiveStreamByIDRequest{StreamID: streamID}
 
@@ -382,7 +382,7 @@ func (s *LiveStreams) parseGetByIDResponse(req *http.Request, httpRes *http.Resp
 //
 // Provide the `streamId` in the request to terminate active connections and remove the stream from the
 // workspace. You can further look for
-// <a href="https://fastpix.com/docs/live-stream-events/live-events#videolive_streamdeleted">video.live_stream.deleted</a>
+// <a href="https://fastpix.com/docs/webhooks/live-events#videolive_streamdeleted">video.live_stream.deleted</a>
 // webhook to notify your system about the status.
 //
 // #### Example
@@ -390,7 +390,7 @@ func (s *LiveStreams) parseGetByIDResponse(req *http.Request, httpRes *http.Resp
 // For an online concert platform, a trial stream was mistakenly made public. The event manager deletes
 // the stream before the concert begins to avoid confusion among viewers.
 //
-// Related guide: <a href="https://fastpix.com/docs/manage-live-streams/create-and-manage-live-streams">Manage streams</a>
+// Related guide: <a href="https://fastpix.com/docs/live-streaming/create-and-manage-live-streams">Manage streams</a>
 func (s *LiveStreams) Delete(ctx context.Context, streamID string, opts ...operations.Option) (*operations.DeleteLiveStreamResponse, error) {
 	request := operations.DeleteLiveStreamRequest{StreamID: streamID}
 
@@ -477,7 +477,7 @@ func (s *LiveStreams) parseDeleteResponse(req *http.Request, httpRes *http.Respo
 // again. By calling this endpoint, the stream is transitioned to a `disabled` state, ensuring it's
 // permanently stopped (unless re-enabled on a paid plan).
 //
-// Related guide <a href="https://fastpix.com/docs/manage-live-streams/create-and-manage-live-streams">Manage streams</a>
+// Related guide <a href="https://fastpix.com/docs/live-streaming/create-and-manage-live-streams">Manage streams</a>
 func (s *LiveStreams) Disable(ctx context.Context, streamID string, opts ...operations.Option) (*operations.DisableLiveStreamResponse, error) {
 	request := operations.DisableLiveStreamRequest{StreamID: streamID}